Modify all files, and made changes to them
