+++ index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="./index.css" />
-    <title>Awesome Books</title>
+    <title>Awesome book</title>
   </head>
   <body>
     <!-- Header section----------- -->
@@ -46,9 +46,9 @@
         <p>Do You have any question or you just want to say hello</p>
         <p>You can reach out to us !</p>
         <ul class="contact-list">
-          <li>Our e-mail : jodavid444@gmail.com</li>
-          <li>Our phone number: +2347014885007</li>
-          <li>Our address: @microvers</li>
+          <li>E-mail : jodavid444dave@gmail.com</li>
+          <li>Phone number: +2347014885007</li>
+          <li>Address: @microverse</li>
         </ul>
       </section>
     </main>
