--- modules/arrayNav.js
@@ -1,16 +1,0 @@
-const list = [
-    {
-      linkId: 'list',
-      pageId: 'listpage',
-    },
-    {
-      linkId: 'add-new',
-      pageId: 'addpage',
-    },
-    {
-      linkId: 'contact-page',
-      pageId: 'contactpage',
-    },
-  ];
-  
-  export default list;
