+++ modules/navArr.js
@@ -1,0 +1,16 @@
+const list = [
+  {
+    linkId: 'list',
+    pageId: 'listpage',
+  },
+  {
+    linkId: 'add-new',
+    pageId: 'addpage',
+  },
+  {
+    linkId: 'contact-page',
+    pageId: 'contactpage',
+  },
+];
+
+export default list;
